Modify test cases
Modified CAPFuntionalTest to use the default values pf AChChargingAddress in AChBillingChargingCharacteristics and TimeDuration objects
Modified ApplyChargingRequestTest to use the default values of the partyToCharge
 and AChChargingAddresses fields
In the AudibleIndicatorTest a new test case is added to both the decode and the encode test methods
In the CAMELAChBillingChargingCharacteristicsTest a new test case is addded to bothe the decode and encode methods and fixed the encoding/decoding of the AudibleIndicator field.
In the testDecode method of TimeDurationChargingResultTest fixed the handling of AChChargingAddress field.